refactor: extract CachePruner and CacheSearch from LocalCache; move Dao out of ui
Two read/reclaim policy clusters leave the LocalCache god object into
sibling classes in the same package, each taking the cache as its only
constructor dependency so the policies are testable in isolation:

- CachePruner: cleanMemory/cleanObservers, the six prune passes
  (hidden/old/expired/superseded/replies+reactions), and the shared
  unlinkAndRemove removal primitive (with removeIfWrap and
  editedTargetIdOf). LocalCache.deleteNote and
  DecryptAndIndexProcessor now call pruner.unlinkAndRemove;
  MemoryTrimmingService drives cache.pruner.*.
  refreshDeletedNoteObservers becomes internal so the pruner can
  notify observers.

- CacheSearch: findUsersStartingWith(username, account),
  findNotesStartingWith, and the three channel prefix searches, plus
  their private exclusion rules. Callers (SearchBarViewModel,
  AgentAttestationScreen, UserSuggestionState, BuzzNewDmViewModel) use
  cache.search.* directly - no delegating shims left behind.

Also moves the Dao interface out of ui/actions/NewMessageTagger.kt into
the model package where its implementor (LocalCache) and its types
live, removing a model-layer interface defined in a UI file.

All moved code is unchanged except for cache. qualification; behavior
is identical. LocalCache.kt: 4554 -> 3921 lines.
